+++ NEWS.md
@@ -17,6 +17,8 @@ fix for the LaTeX problem sanitized with the last `rticles` update (thanks, @RLu
 
 - Update all templates regarding CSLReference environment following changes in Pandoc's default template.
 
+- Add article template `pihph_article()` for the *Papers in Historical Phonology* (PiHPh).
+
 rticles 0.17
 ---------------------------------------------------------------------